Add exports, DI architecture, algorithm fixes, and a cross-platform CLI
A broad set of improvements across the library, sample, and a new test suite.

Features
- Palette export system: IPaletteExporter + PaletteExporter facade with six
  formats (JSON, hex list, C/Arduino array, CSS, SCSS, Tailwind), plus a
  byte[] extraction overload for already-in-memory images.
- Image-format validation: magic-byte sniffing rejects non-images before the
  decoder runs (UnsupportedImageFormatException).
- maxDegreeOfParallelism parameter with deterministic parallel K-Means.
- PaletteQuality.MeanDeltaE objective quality metric.

Architecture
- ColorPaletteExtractorBase template + per-algorithm Options records.
- PaletteGenerator/PaletteExporter take IEnumerable<> for DI, with parameterless
  convenience ctors; reflection discovery removed.
- New OverTone.Extensions.DependencyInjection package (AddOverTone()); the core
  OverTone package stays dependency-free.

Algorithm correctness
- Wu: fixed the 3-D cumulative moments (green axis was never integrated) and the
  int overflow in split scoring; Wu now produces correct palettes.
- K-Means: k-means++ seeding + fixed-seed determinism.
- Octree: fixed byte-overflow when requesting > 255 colors.
- Fuzzy C-Means: deterministic; Median Cut + FCM now subsample large images.
- SelectDiverse: no longer emits duplicate / 0-pixel colors.

Sample (now cross-platform)
- Retargeted to net10.0 (dropped WinForms); console path prompt + CLI argument;
  session loop; compare-algorithms view ranked by mean delta-E; richer
  swatch/RGB/HSL display.

Tests & docs
- New OverTone.Tests (xUnit): 46 tests across exporters, algorithm robustness,
  DI, validation, and parallelism.
- README + TODO updated: selection modes, tuning, large images/RAW handling,
  and an expanded algorithm roadmap.

+++ OverTone/Algorithms/Octree.cs
@@ -38,7 +38,7 @@ public class Octree
     /// to <paramref name="maxColors"/>. Nodes are merged from upper levels downward.
     /// </summary>
     /// <param name="maxColors">Maximum number of leaf nodes to retain.</param>
-    public void Reduce(byte maxColors)
+    public void Reduce(int maxColors)
     {
         // Merge nodes until leaf count <= maxColors
         var leaves = GetLeafCount();
